// src/app/sections/models/clinic.model.ts
export class Clinic
  {
    id: number;
    orgbet: string;
    klartext: string;
    avdTeam: string;
    avdText: string;
    vrdkod: string;
    sjukhus: string;
    huskropp: string;
    ansvar: string;
    databasId: number;
    verkOmr: string;
    orgbetRemklin: string;
    nivaa: string;
    avdService: string;
    division: string;
    orgbetOver: string;
    irkNr: string;
    orgbetSvarsklin: string;
    orgbetDebklin: string;
    hsaId: string;
    eremiss: number;
    fomDat: Date;
    tomDat: Date;
    updtid: Date;
  }